+++ projects/angular-sdk/src/components/cfar-offer-banner-large/cfar-offer-banner-large.component.ts
@@ -26,6 +26,8 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
 
   @Output() chooseCoverage = new EventEmitter();
 
+  private contractsByChoiceIndex = new Map<number, CfarContractCustomer>();
+
   constructor(
     private _adapter: DateAdapter<any>,
     private _translateService: TranslateService,
@@ -46,7 +48,7 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
       this.isLoading = true;
 
       this._hopperProxyService
-        .postCfarOffers(this.basePath, this.hCSessionId, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarOfferRequest()))
+        .postCfarOffers(this.basePath, this.hCSessionId, this.currentLang, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarOfferRequest()))
         .pipe(take(1))
         .subscribe(
           (cfarOffers) => {
@@ -94,23 +96,32 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
       if (this.isFakeBackend) {
         this.chooseCoverage.emit(this._buildFakePostCfarContractsResponse());
       } else {
-        this.isLoading = true;
+        // Emit the cached element
+        if (this.contractsByChoiceIndex.has(this.selectedChoice)) {
+          this.chooseCoverage.emit(this.contractsByChoiceIndex.get(this.selectedChoice));
+        // Backend call
+        } else {
+          this.isLoading = true;
 
-        // Create CFAR Contract
-        this._hopperProxyService
-          .postCfarContracts(this.basePath, this.hCSessionId, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarContractRequest()))
-          .pipe(take(1))
-          .subscribe(
-            (cfarContract: CfarContractCustomer) => {
-              this.chooseCoverage.emit(cfarContract);
-              this.isLoading = false;
-            },
-            (error) => {
-              console.error(error);
-              this.isLoading = false;
-            }
-          );
-      } 
+          // Create CFAR Contract
+          this._hopperProxyService
+            .postCfarContracts(this.basePath, this.hCSessionId, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarContractRequest()))
+            .pipe(take(1))
+            .subscribe(
+              (cfarContract: CfarContractCustomer) => {
+                // Cache the result into a map (for API performance)
+                this.contractsByChoiceIndex.set(this.selectedChoice, cfarContract);
+  
+                this.chooseCoverage.emit(cfarContract);
+                this.isLoading = false;
+              },
+              (error) => {
+                console.error(error);
+                this.isLoading = false;
+              }
+            );
+        } 
+      }
     } else {
       this.chooseCoverage.emit(null);
     }
@@ -148,8 +159,8 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
     return [
       {
         id: "1ecf859e-8785-625f-8eda-198d1ce0d6c4",
-        premium: "861.00",
-        coverage: "5736.78",
+        premium: "8.00",
+        coverage: "57.78",
         currency: "CAD",
         requestType: "ancillary",
         toUsdExchangeRate: "0.7744877537996369201410187302118379",
@@ -213,7 +224,7 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
             }
           ],
           ancillaries: [],
-          totalPrice: "7170.96"
+          totalPrice: "71.96"
         },
         offerDescription: [
           "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",
@@ -223,8 +234,8 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
       },
       {
         id: "1ecf859e-8785-625f-8eda-198d1ce0d6c5",
-        premium: "1076.00",
-        coverage: "7170.96",
+        premium: "10.00",
+        coverage: "71.96",
         currency: "CAD",
         requestType: "ancillary",
         toUsdExchangeRate: "0.7744877537996369201410187302118379",
@@ -288,7 +299,7 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
             }
           ],
           ancillaries: [],
-          totalPrice: "7170.96"
+          totalPrice: "71.96"
         },
         offerDescription: [
           "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",