+++ projects/angular-sdk/src/components/cfar-offer-banner/cfar-offer-banner.component.ts
@@ -49,7 +49,7 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
       this.isLoading = true;
 
       this._hopperProxyService
-        .postCfarOffers(this.basePath, this.hCSessionId, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarOfferRequest()))
+        .postCfarOffers(this.basePath, this.hCSessionId, this.currentLang, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarOfferRequest()))
         .pipe(take(1))
         .subscribe(
           (cfarOffers) => {
@@ -116,8 +116,8 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
     return [
       {
         id: "1ecf859e-8785-625f-8eda-198d1ce0d6c4",
-        premium: "861.00",
-        coverage: "5736.78",
+        premium: "8.00",
+        coverage: "57.78",
         currency: "CAD",
         requestType: "ancillary",
         toUsdExchangeRate: "0.7744877537996369201410187302118379",
@@ -181,7 +181,7 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
             }
           ],
           ancillaries: [],
-          totalPrice: "7170.96"
+          totalPrice: "71.96"
         },
         offerDescription: [
           "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",
@@ -191,8 +191,8 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
       },
       {
         id: "1ecf859e-8785-625f-8eda-198d1ce0d6c5",
-        premium: "1076.00",
-        coverage: "7170.96",
+        premium: "10.00",
+        coverage: "71.96",
         currency: "CAD",
         requestType: "ancillary",
         toUsdExchangeRate: "0.7744877537996369201410187302118379",
@@ -256,7 +256,7 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
             }
           ],
           ancillaries: [],
-          totalPrice: "7170.96"
+          totalPrice: "71.96"
         },
         offerDescription: [
           "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",