+++ projects/angular-sdk/src/components/cfar-offer-dialog/cfar-offer-dialog.component.ts
@@ -62,7 +62,7 @@ export class CfarOfferDialogComponent extends GlobalComponent implements OnInit,
       this.isLoading = true;
   
       this._hopperProxyService
-        .postCfarOffers(this.basePath, this._hCSessionId, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarOfferRequest()))
+        .postCfarOffers(this.basePath, this._hCSessionId, this.currentLang, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarOfferRequest()))
         .pipe(take(1))
         .subscribe(
           (cfarOffers) => {
@@ -150,8 +150,8 @@ export class CfarOfferDialogComponent extends GlobalComponent implements OnInit,
     return [
       {
         id: "1ecf859e-8785-625f-8eda-198d1ce0d6c4",
-        premium: "861.00",
-        coverage: "5736.78",
+        premium: "8.00",
+        coverage: "57.78",
         currency: "CAD",
         requestType: "ancillary",
         toUsdExchangeRate: "0.7744877537996369201410187302118379",
@@ -215,7 +215,7 @@ export class CfarOfferDialogComponent extends GlobalComponent implements OnInit,
             }
           ],
           ancillaries: [],
-          totalPrice: "7170.96"
+          totalPrice: "71.96"
         },
         offerDescription: [
           "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",
@@ -225,8 +225,8 @@ export class CfarOfferDialogComponent extends GlobalComponent implements OnInit,
       },
       {
         id: "1ecf859e-8785-625f-8eda-198d1ce0d6c5",
-        premium: "1076.00",
-        coverage: "7170.96",
+        premium: "10.00",
+        coverage: "71.96",
         currency: "CAD",
         requestType: "ancillary",
         toUsdExchangeRate: "0.7744877537996369201410187302118379",
@@ -290,7 +290,7 @@ export class CfarOfferDialogComponent extends GlobalComponent implements OnInit,
             }
           ],
           ancillaries: [],
-          totalPrice: "7170.96"
+          totalPrice: "71.96"
         },
         offerDescription: [
           "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",