+++ src/app/pages/guides/cfar-exercise-dialog/cfar-exercise-dialog.component.ts
@@ -99,7 +99,7 @@ export class CfarExerciseDialogPageComponent extends CommonGuidesComponent {
         name: 'currentTheme',
         description: `
           The active theme managed by Hopper.
-          If you have an Angular material theme, this field can be removed
+          If you have an Angular material theme, this field can be ignored
         `,
         required: false
       },